На прорверку(от 06.04) задания 3.4 "Введение в SQL" добавлен swagger Исправлено: -удалена зависимость oracle -запросы sql запитаны полностью -добавлена аннотация @JsonIgnore -поля класса расположены на уровне экземплярных полей

+++ src/main/java/ru/hogwarts/school/repository/StudentRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface StudentRepository extends JpaRepository <Student, Long> {
     List <Student> findByAge (Long age);
-    List<Student>findByAgeBetween(Long ageMin ,Long ageMax);
+    List<Student>findByAgeBetween(int ageMin ,int ageMax);
 }
